reactor(test): mock mock data to separate file
- move server side test mock data to dummyData.json file
- move client side test mock data to samples.js, componentMock.js and action.js
- split action and reducers test to different files

[Delivers #152410973]
