+++ client/__tests__/actions/signupActions.test.js
@@ -1,0 +1,35 @@
+/* global expect */
+import configureMockStore from 'redux-mock-store';
+import thunk from 'redux-thunk';
+import axios from 'axios';
+import MockAdapter from 'axios-mock-adapter';
+import samples from '../../__mocks__/samples';
+import userSignupRequest from '../../actions/signupActions';
+import * as types from '../../actions/types';
+
+const middlewares = [thunk];
+const mockStore = configureMockStore(middlewares);
+const mock = new MockAdapter(axios);
+
+describe('Signup Action', () => {
+  const user = samples.user;
+  const token = samples.validToken;
+
+  const store = mockStore({});
+  const userData = samples.userData;
+
+  it('should have userSignupRequest action return json repsonse', () => {
+    mock.onPost('/api/v1/users/signup', userData)
+      .reply(200, { token,
+        message: 'User created successfully',
+        username: 'Kene',
+        success: true });
+    return store.dispatch(userSignupRequest(userData))
+      .then(() => {
+        expect(store.getActions()).toEqual([{
+          type: types.SET_CURRENT_USER,
+          user
+        }]);
+      });
+  });
+});
