+++ client/__tests__/components/ChangePassword.test.jsx
@@ -1,4 +1,3 @@
-// import
 /* global expect */
 import React from 'react';
 import sinon from 'sinon';
@@ -16,7 +15,7 @@ describe('<ForgotPassword />', () => {
     .resolve({ data: { message: 'Success' }, response: { data: 'Error' } }));
   const forgotPasswordRequest = sinon.spy(() => Promise
     .resolve({ data: { message: 'Success' }, response: { data: 'Error' } }));
-  // assign all props to a varibale props
+
   const props = {
     params,
     isValidToken,
