+++ client/__tests__/components/CreateGroup.test.jsx
@@ -1,4 +1,3 @@
-// import
 /* global expect */
 import React from 'react';
 import { shallow } from 'enzyme';
