+++ server/app.js
@@ -1,4 +1,3 @@
-// *** main dependencies *** //
 import express from 'express';
 import flash from 'express-flash';
 import path from 'path';
@@ -10,28 +9,21 @@ import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
 import routes from './routes';
 
-// *** express instance *** //
+const config = process.env.NODE_ENV !== 'production' ?
+  require('../webpack.config') : '';
+
 const app = express();
 
-// *** config middleware *** //
 app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-
-// *** main routes *** //
 app.use('/', routes);
-
-// use express flash message
 app.use(flash());
 
 if (process.env.NODE_ENV !== 'production') {
-  const config = require('../webpack.config');
-
-  // *** webpack compiler ***
   const compiler = webpack(config);
 
-  // *** webpack middleware
   app.use(webpackDevMiddleware(compiler, {
     noInfo: true,
     publicPath: config.output.publicPath
@@ -58,15 +50,12 @@ if (process.env.NODE_ENV !== 'production') {
   });
 }
 
-
-// catch 404 and forward to error handler
 app.use((req, res, next) => {
   const err = new Error('Not Found');
   err.status = 404;
   next(err);
 });
 
-// Listen
 app.listen(process.env.PORT || 3000);
 
 export default app;
