--- server/config.js
@@ -1,3 +1,0 @@
-export default {
-  jwtSecret: 'mynameiskenechukwu' // process.env.JWT_SECRET
-};
