+++ server/controller/user.js
@@ -2,9 +2,10 @@ import jwt from 'jsonwebtoken';
 import md5 from 'md5';
 import crypto from 'crypto';
 import { Users } from '../models';
-import config from '../config';
 import sendMail from '../utils/sendMail';
 import validateInput from '../shared/validations/validateInput';
+
+require('dotenv').config();
 
 export default {
   /**
@@ -59,7 +60,7 @@ export default {
               const token = jwt.sign({
                 id: aUser.id,
                 username: aUser.username
-              }, config.jwtSecret, { expiresIn: '48h' }); // expires in 48h
+              }, process.env.JWT_SECRET, { expiresIn: '48h' });
               res.status(201).json({
                 token,
                 message: 'User Created successfully',
@@ -109,7 +110,7 @@ export default {
             const token = jwt.sign({
               id: user[0].id,
               username: user[0].username
-            }, config.jwtSecret, { expiresIn: '48h' }); // token expires in 48h
+            }, process.env.JWT_SECRET, { expiresIn: '48h' });
 
             res
               .status(202)
@@ -125,8 +126,9 @@ export default {
             .send({
               message: 'Username not found, please register'
             });
-        }).catch(() => {
+        }).catch((error) => {
           res.status(500).json({
+            error: error.errors.message,
             message: 'An error has occurred trying to search for user',
           });
         });
@@ -157,7 +159,7 @@ export default {
         .then((user) => {
           if (user[0]) {
             const token = crypto.randomBytes(20).toString('hex');
-            const tokenExpireDate = Date.now() + 3600000; // expire in 1hr
+            const tokenExpireDate = Date.now() + 3600000;
             Users.update({
               resetPasswordToken: token,
               resetPasswordExpiryTime: tokenExpireDate,
@@ -228,8 +230,8 @@ export default {
     } else {
       return Users.update({
         password: md5(req.body.password),
-        resetPasswordToken: '', // resetPasswordToken set to empty
-        resetPasswordExpiryTime: '' // resetPasswordExpiryTime set to empty
+        resetPasswordToken: '',
+        resetPasswordExpiryTime: ''
       }, {
         where: {
           resetPasswordToken: req.params.token
