+++ server/shared/middleware/jwtAuth.js
@@ -1,6 +1,4 @@
-
 import jwt from 'jsonwebtoken';
-import config from '../../config';
 
 /**
  * Middleware for protecting router using jwt
@@ -10,18 +8,15 @@ import config from '../../config';
  * @return {json} returns json object as a response
  */
 export default (req, res, next) => {
-  // let token;
   const token = req.body.token || req.query.token ||
         req.headers.Authorization || req.headers['x-access-token'];
   if (token) {
-    jwt.verify(token, config.jwtSecret, (err, authToken) => {
+    jwt.verify(token, process.env.JWT_SECRET, (err, authToken) => {
       if (err) {
         res.status(401).send({
           message: 'Sorry, user not authenticated, invalid access token'
         });
       } else {
-        // if authenticated with auth token,
-        // save auth token and to request for use in other routes
         req.authToken = authToken;
         authToken = JSON.stringify(authToken);
         res.append('user', authToken);
