+++ server/test/routes/groupControllerTest.js
@@ -1,11 +1,10 @@
-// Require the dev-dependencies
 import supertest from 'supertest';
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import { Groups, userGroups } from '../models';
-import app from '../app';
+import dummyData from '../dummy.json';
+import { Groups, userGroups } from '../../models';
+import app from '../../app';
 
-// During the test the env variable is set to test
 process.env.NODE_ENV = 'test';
 
 const server = supertest.agent(app);
@@ -13,14 +12,16 @@ const should = chai.should();
 
 chai.use(chaiHttp);
 
+const { group } = dummyData.Groups;
+const { validSigninAccount } = dummyData.Users;
+
+
 describe('Groups', () => {
   let token = '';
-  before((done) => { // Before each test we empty the database
+  before((done) => {
     server
       .post('/api/v1/users/signin')
-      .send({
-        username: 'kene',
-        password: 'kene' })
+      .send(validSigninAccount)
       .end((err, res) => {
         token = res.body.token;
       });
@@ -34,17 +35,16 @@ describe('Groups', () => {
       });
   });
 
-  describe('Controller Tests: ', () => {
-    // Test for succesful group creation
-    describe('Group', () => {
+  describe('API Routes Test: ', () => {
+    describe('/api/v1/groups', () => {
       it('should create a group when all parameters are supplied',
         (done) => {
           server
             .post('/api/v1/groups')
             .send({
               token,
-              groupName: 'Group',
-              createdby: 1,
+              groupName: group.groupName,
+              createdby: group.createdby,
             })
             .end((err, res) => {
               res.should.have.status(201);
@@ -62,8 +62,8 @@ describe('Groups', () => {
             .post('/api/v1/groups')
             .send({
               token,
-              groupName: '',
-              createdby: 1,
+              groupName: group.emptyGroupName,
+              createdby: group.createdby,
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -79,9 +79,10 @@ describe('Groups', () => {
           server
             .post('/api/v1/groups')
             .send({
+
+              // assuming createdby is missing
               token,
-              groupName: 'Group',
-              // createdby: 1, // missing parameter
+              groupName: group.groupName
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -91,7 +92,6 @@ describe('Groups', () => {
               done();
             });
         });
-      // Test for uniquness of a group
       it(`should not create a group with a
           group name that already exists`,
         (done) => {
@@ -99,8 +99,8 @@ describe('Groups', () => {
             .post('/api/v1/groups')
             .send({
               token,
-              groupName: 'Group',
-              createdby: 1,
+              groupName: group.groupName,
+              createdby: group.createdby,
             })
             .end((err, res) => {
               res.should.have.status(409);
@@ -118,7 +118,7 @@ describe('Groups', () => {
             .post('/api/v1/groups/creator')
             .send({
               token,
-              userId: '1'
+              userId: group.user
             })
             .end((err, res) => {
               res.should.have.status(200);
@@ -131,10 +131,9 @@ describe('Groups', () => {
               done();
             });
         });
-      // Test for retrieving group by its id
       it('should return groups by its id',
         (done) => {
-          const groupId = '1';
+          const groupId = group.groupId;
           server
             .get(`/api/v1/groups/${groupId}`)
             .set({ 'x-access-token': token })
@@ -149,7 +148,7 @@ describe('Groups', () => {
               done();
             });
         });
-      it('should return 404 for group not found',
+      it('should return not found for group not found',
         (done) => {
           const groupId = null;
           server
@@ -169,7 +168,7 @@ describe('Groups', () => {
             .post('/api/v1/groups/creator')
             .send({
               token,
-              userId: 3
+              userId: group.invalidUserId
             })
             .end((err, res) => {
               res.should.have.status(500);
@@ -183,8 +182,9 @@ describe('Groups', () => {
           server
             .post('/api/v1/groups/creator')
             .send({
+
+              // userId not passed
               token,
-              // userId: 1 // userId not passed
             })
             .end((err, res) => {
               res.should.have.status(400);
