+++ server/test/routes/groupUserControllerTest.js
@@ -1,26 +1,30 @@
-// Require the dev-dependencies
 import supertest from 'supertest';
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import { Groups, userGroups } from '../models';
-import app from '../app';
+import { userGroups } from '../../models';
+import dummyData from '../dummy.json';
+import app from '../../app';
 
-// During the test the env variable is set to test
 process.env.NODE_ENV = 'test';
 
 const server = supertest.agent(app);
-const should = chai.should();
-
 chai.use(chaiHttp);
+
+const { group } = dummyData.Groups;
+const {
+  validUser,
+  invalidUser,
+  emptyUsername,
+  anotherValidUser,
+  validSigninAccount
+} = dummyData.Users;
 
 describe('GroupsUser', () => {
   let token = '';
-  before((done) => { // Before each test we empty the database
+  before((done) => {
     server
       .post('/api/v1/users/signin')
-      .send({
-        username: 'kene',
-        password: 'kene' })
+      .send(validSigninAccount)
       .end((err, res) => {
         token = res.body.token;
       });
@@ -31,17 +35,16 @@ describe('GroupsUser', () => {
       });
   });
 
-  describe('Controller Tests: ', () => {
-    // Test for succesful group creation
-    describe('Group', () => {
-      const groupId = 1;
+  describe('API Routes Test: ', () => {
+    describe('/api/v1/groups/:groupId/user', () => {
+      const groupId = group.groupId;
       it('should add a user to a group',
         (done) => {
           server
             .post(`/api/v1/groups/${groupId}/user`)
             .send({
               token,
-              userId: '1'
+              userId: validUser.userId
             })
             .end((err, res) => {
               res.should.have.status(201);
@@ -59,7 +62,7 @@ describe('GroupsUser', () => {
             .post(`/api/v1/groups/${groupId}/user`)
             .send({
               token,
-              userId: '2'
+              userId: anotherValidUser.userId
             })
             .end((err, res) => {
               res.should.have.status(201);
@@ -76,8 +79,9 @@ describe('GroupsUser', () => {
           server
             .post(`/api/v1/groups/${groupId}/user`)
             .send({
+
+              // assuming userId is not supplied
               token,
-              // userId: '1'
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -93,7 +97,7 @@ describe('GroupsUser', () => {
             .post(`/api/v1/groups/${groupId}/user`)
             .send({
               token,
-              userId: '100' // this userId does not exist
+              userId: invalidUser.userId
             })
             .end((err, res) => {
               res.should.have.status(404);
@@ -109,7 +113,7 @@ describe('GroupsUser', () => {
             .post(`/api/v1/groups/${groupId}/user`)
             .send({
               token,
-              userId: '1' // this userId already exist in a group
+              userId: validUser.userId
             })
             .end((err, res) => {
               res.should.have.status(409);
@@ -125,7 +129,7 @@ describe('GroupsUser', () => {
             .post(`/api/v1/groups/${undefined}/user`)
             .send({
               token,
-              userId: '1'
+              userId: validUser.userId
             })
             .end((err, res) => {
               res.should.have.status(500);
@@ -141,9 +145,10 @@ describe('GroupsUser', () => {
           server
             .post(`/api/v1/groups/${groupId}/users`)
             .send({
+
+              // assuming userId is missing
               token,
-              admin: 1,
-              // user: 6 // user is missing
+              admin: group.admin
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -160,8 +165,8 @@ describe('GroupsUser', () => {
             .post(`/api/v1/groups/${groupId}/users`)
             .send({
               token,
-              admin: '1',
-              user: '1',
+              admin: group.admin,
+              user: validUser.userId,
             })
             .end((err, res) => {
               res.should.have.status(401);
@@ -177,8 +182,8 @@ describe('GroupsUser', () => {
             .post(`/api/v1/groups/${groupId}/users`)
             .send({
               token,
-              admin: '1',
-              user: '100', // this user does not exist/belong to group 1
+              admin: group.admin,
+              user: invalidUser.userId,
             })
             .end((err, res) => {
               res.should.have.status(404);
@@ -194,8 +199,8 @@ describe('GroupsUser', () => {
             .post(`/api/v1/groups/${groupId}/users`)
             .send({
               token,
-              admin: '1',
-              user: '2',
+              admin: group.admin,
+              user: anotherValidUser.userId,
             })
             .end((err, res) => {
               res.should.have.status(200);
@@ -213,7 +218,7 @@ describe('GroupsUser', () => {
             .post('/api/v1/user/groups')
             .send({
               token,
-              username: 'kene',
+              username: validUser.username,
             })
             .end((err, res) => {
               res.should.have.status(200);
@@ -231,8 +236,9 @@ describe('GroupsUser', () => {
           server
             .post('/api/v1/user/groups')
             .send({
+
+              // assuming the username is not supplied
               token,
-              // username: 'kene',
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -249,7 +255,7 @@ describe('GroupsUser', () => {
             .post('/api/v1/user/groups')
             .send({
               token,
-              username: '',
+              username: emptyUsername.username,
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -266,13 +272,13 @@ describe('GroupsUser', () => {
             .post('/api/v1/users/0')
             .send({
               token,
-              username: 'k',
+              username: validUser.username,
             })
             .end((err, res) => {
               res.should.have.status(200);
               res.body.should.be.a('object');
               res.body.should.have.property('count')
-                .eql(2);
+                .eql(1);
               res.body.should.have.property('rows');
               done();
             });
@@ -282,8 +288,9 @@ describe('GroupsUser', () => {
           server
             .post('/api/v1/users/0')
             .send({
-              token,
-              // username: 'k',
+
+              // assuming username is not supplied
+              token
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -298,7 +305,7 @@ describe('GroupsUser', () => {
             .post('/api/v1/users/0')
             .send({
               token,
-              username: '',
+              username: emptyUsername.username,
             })
             .end((err, res) => {
               res.should.have.status(400);
@@ -307,19 +314,22 @@ describe('GroupsUser', () => {
               done();
             });
         });
-      
-      it(`should unauthorized status code when trying 
-        calling the endpoint without toke`,
+
+      it(`should throw unauthorized error when trying 
+        calling the groupUser endpoint without token`,
         (done) => {
           server
             .post('/api/v1/users/0')
             .send({
-              // token, // token is not available
-              username: '3',
+
+              // token is not available
+              username: validUser.username,
             })
             .end((err, res) => {
               res.should.have.status(403);
               res.body.should.be.a('object');
+              res.body.should.have.property('message')
+                .eql('No token provided.');
               done();
             });
         });
