+++ server/utils/sendMail.js
@@ -1,16 +1,17 @@
 import nodemailer from 'nodemailer';
+
 /**
  * sendMail - This function send Mail notifcation for PostIT
- * @param  {array} receivers An array of users
- * the group
+ * @param  {array} receivers An array of email receivers
  * @param  {string} message message to be sent
  * @param {string} subject title of the message to be sent
- * @return {boolean} returns true or false                  
+ * @return {boolean} returns true or false
  */
 export default function sendMail(receivers, message, subject) {
   let emailReceivers = '';
   receivers.map((user) => {
     emailReceivers = `${emailReceivers + user.email},`;
+    return emailReceivers;
   });
 
   // create reusable transporter object using the default SMTP transport
@@ -24,9 +25,9 @@ export default function sendMail(receivers, message, subject) {
 
   // setup email data with unicode symbols
   const mailOptions = {
-    from: process.env.EMAIL_NAME, // sender address
-    to: emailReceivers, // list of receivers
-    subject, // Subject line
+    from: process.env.EMAIL_NAME,
+    to: emailReceivers,
+    subject,
     html: message
   };
