+++ webpack.config.prod.js
@@ -17,7 +17,6 @@ module.exports = {
   resolve: {
     extensions: ['*', '.js', '.jsx']
   },
-  // define the module loaders to be used by the webpack
   module: {
     loaders: [
       {
@@ -45,14 +44,11 @@ module.exports = {
       },
     ]
   },
-
-  // specify the output path for bundled file
   output: {
     path: path.resolve('./client/dist'),
     filename: 'bundle.min.js',
     publicPath: '/dist/'
   },
-  // // assign the node files that cannot be found to empty
   node: {
     net: 'empty',
     dns: 'empty',
@@ -60,7 +56,6 @@ module.exports = {
     tls: 'empty',
     child_process: 'empty',
   },
-  // webpack plugins
   plugins: [
     new HtmlWebpackPlugin({
       title: 'Production',
@@ -75,11 +70,6 @@ module.exports = {
     }),
     new webpack.optimize.OccurrenceOrderPlugin(),
     new webpack.NoEmitOnErrorsPlugin(),
-    new webpack.DefinePlugin({
-      'process.env': {
-        NODE_ENV: JSON.stringify('production')
-      }
-    }),
     new webpack.optimize.UglifyJsPlugin(),
     new webpack.DefinePlugin({
       'process.env': {
@@ -87,7 +77,8 @@ module.exports = {
         SMS_API_KEY: JSON.stringify(process.env.SMS_API_KEY),
         SMS_API_SECRET: JSON.stringify(process.env.SMS_API_SECRET),
         EMAIL_NAME: JSON.stringify(process.env.EMAIL_NAME),
-        EMAIL_PASSWORD: JSON.stringify(process.env.EMAIL_PASSWORD)
+        EMAIL_PASSWORD: JSON.stringify(process.env.EMAIL_PASSWORD),
+        JWT_SECRET: JSON.stringify(process.env.JWT_SECRET)
       },
     })
   ],
